demo_code/mcp_entra/requirements.txt
# FastMCP with Azure Entra OAuth dependencies
fastmcp>=2.12.0
python-dotenv>=1.0.0
httpx>=0.24.0

# Optional: Additional dependencies for enhanced functionality
requests>=2.31.0
pydantic>=2.0.0